Update assignment.html
Previously upon clicking the link, it was trying to go to index.html within the current directory, which is IIITHResources/CPro/Web (I just used shortforms to representthe actual name.) And this page is nonexistent, so it was showingthe error page. But it should actually go to the main homepage, which is two directories up. So I added in the changes for that.